implement a processor utility to facilitate manual evaluation
this utility reads the tsv and spawn a vim instance containing the
original method commented on top of the file, and the partially
detokenized method.
if a modification occurs, the modified method will be saved in the
"detokenized_method" column. if not, the method will be skipped.
To exit the script the user must delete everything in the temp file
and save. The script will apply the changes and stop. Next time, you
will start from the method you deleted to previous time