change toCamelCase implementation
I changed the implementation, as I realized that a string could have
trailing characters in the center too. I therefore decided to use a binary
tree to parse the strings and apply the camelCase function only
on strings that are not trailing characters. Then, using a
bit of recursion i reconstructed the original string